// apps/client/src/components/types.ts
export type NavigationButtonType =
	| "Home"
	| "Settings"
	| "Money"
	| "Message"
	| "Contacts"
	| "Clock"
	| "Shop"
	| "Cart"
	| "Menu"
	| "Search"
	| "Tool"
	| "UserAdd"
	| "Plus"
	| "Close"
	| "ListStar"
	| "Bag"
	| "CreditCard"
	| "RadioGrid"
	| "Payment"
	| "Radio"
	| "Logout"
	| "ArrowBack"
	| "Table"
	| "Fork"
	| "Dashboard"
	| "RaisedHand"
	| "TrashCan"
	| "QRCode"
	| "People"
	| "Phone"
	| "StarFill"
	| "StarOutline"
	| "ClipBoard"
	| "ArrowDown"
	| "ArrowUp"
	| "Location"
	| "Calendar";